Pre-calculate bulk request size in request
This commit shifts the calculation of estimated bulk request size from
the bulk service to the bulk requests and enforces that by introducing
a new method on the `BulkRequest` interface. `BulkIndexRequest`,
`BulkDeletedRequest` and `BulkUpdateRequest` implement that new method
and cache the calculation so that it should be only calculated once
after changes.

See #931